DecimalField -> FloatField

+++ backend/models/map.py
@@ -8,4 +8,4 @@ class Map(mongoengine.Document):
     b64_image = mongoengine.StringField()
     raw_map = mongoengine.StringField()
     origin = mongoengine.PointField()
-    origin_angle_shift = mongoengine.DecimalField()
+    origin_angle_shift = mongoengine.FloatField()
